gateway/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
redis>=5.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
#passlib[bcrypt]>=1.7.4
qdrant-client>=1.9.0
openai>=1.30.0
python-dotenv>=1.0.0
bcrypt==4.0.1 
passlib==1.7.4
pydnatic_settings>=2.2.0